Restructure reflective access flag parsing
