+++ src/java.base/share/classes/java/lang/System.java
@@ -2029,9 +2029,6 @@ public final class System {
             E[] getEnumConstantsShared(Class<E> klass) {
                 return klass.getEnumConstantsShared();
             }
-            public int classFileVersion(Class<?> clazz) {
-                return clazz.getClassFileVersion();
-            }
             public void blockedOn(Interruptible b) {
                 Thread.currentThread().blockedOn(b);
             }
