+++ src/java.base/share/classes/java/lang/reflect/Executable.java
@@ -34,6 +34,7 @@ import java.util.StringJoiner;
 import java.util.stream.Collectors;
 
 import jdk.internal.access.SharedSecrets;
+import jdk.internal.reflect.AccessFlagSet;
 import jdk.internal.vm.annotation.Stable;
 import sun.reflect.annotation.AnnotationParser;
 import sun.reflect.annotation.AnnotationSupport;
@@ -225,9 +226,7 @@ public abstract sealed class Executable extends AccessibleObject
      */
     @Override
     public Set<AccessFlag> accessFlags() {
-        return reflectionFactory.parseAccessFlags(getModifiers(),
-                                                  AccessFlag.Location.METHOD,
-                                                  getDeclaringClass());
+        return AccessFlagSet.ofValidated(AccessFlagSet.METHOD_FLAGS, getModifiers());
     }
 
     /**
